Feature/exeption

+++ Deleter.cpp
@@ -1,21 +1,21 @@
 #include "Deleter.h"
 
-void Deleter::del(path path, vector<string>& ext) {
+void Deleter::del(path path, vector<string>& ext, vector<string>& exeptions) {
 	if (exists(path)) {
-		if (checker(path.filename().string(), ext)) {
+		if (!checker(path.filename().string(), exeptions) && checker(path.filename().string(), ext)) {
 			if (is_directory(path)) remove_all(path);
 			else remove(path);
 			return;
 		}
 		for (auto& it : directory_iterator(path)) {
 			if (is_directory(it.path())) {
-				if (checker(it.path().filename().string(), ext)) {
+				if (!checker(it.path().filename().string(), exeptions) && checker(it.path().filename().string(), ext)) {
 					remove_all(it.path());
 				}
-				else del(it.path(), ext);
+				else del(it.path(), ext, exeptions);
 			}
 			else {
-				if (checker(it.path().filename().string(), ext)) {
+				if (!checker(it.path().filename().string(), exeptions) && checker(it.path().filename().string(), ext)) {
 					remove(it.path());
 				}
 			}
@@ -27,6 +27,7 @@ void Deleter::del(path path, vector<string>& ext) {
 }
 
 bool Deleter::checker(string name, vector<string>& del_list) {
+	//тут ошибка
 	for (int i = 0; i < del_list.size(); i++) {
 		int left = 0, right = 0;
 		int l = 0, rs = 0;
@@ -71,17 +72,30 @@ void Deleter::ui_asking() {
 		string com;
 		getline(cin, com);
 		if (com == "1") {
-			string path, d;
+			string path, d = "";
 			vector<string> del_vec;
+			vector<string> exeptions;
 			cout << "Введите путь: ";
 			getline(cin, path);
 			if (exists(path)) {
+				//тут ошибка
 				cout << "Введите список ключевых слов для удаления (если список закончен, то введите '.'): " << endl;
 				do {
-					getline(cin, d);
-					del_vec.push_back(d);
+					if (d != ".") {
+						getline(cin, d);
+						del_vec.push_back(d);
+					}
 				} while (d != ".");
-				del(path, del_vec);
+				d = "";
+				//del(path, del_vec);
+				cout << "Введите список ключевых слов для исключений, это файлы, которые не будут удалены (если список закончен, то введите '.'): ";
+				do {
+					if (d != ".") {
+						getline(cin, d);
+						exeptions.push_back(d);
+					}
+				} while (d != ".");
+				del(path, del_vec, exeptions);
 			}
 			else {
 				cout << "Такого пути не существует" << endl;
